refactor: move S3 helper functions and constants to s3Client module
